+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Please mark all change in change log and use the issue from GitHub
 ## Feature
 - \#343 - Add Opentracing
 - \#665 - Support get/set config via CLI
+- \#771 - Add server build commit info interface
 
 ## Improvement